[REVERTME] DollarCove TB: Set Vnn VSLEEP to 0.5V for TI PMIC
BZ: 206306

If Vsleep Vnn=0.45V, WLAN cannot function.
Set Vnn to 0.5V during S3, so that WLAN functionality is not broken.
This is done from OS as the BIOS cannot be changed at this stage
due to compatibility issues with Windows.

This patch should be reverted once the issue is resolved by the H/W
or connectivity team.

Change-Id: If5076a31debd71ad747a28641089378752d4fed0
Signed-off-by: Souvik Kumar Chakravarty <souvik.k.chakravarty@intel.com>