Telephony/HSI: map TRACE channel by default to EDLP channel 4
In current code, by default, EDLP channel 4 was mapped to RMNET2
link (that is never used). This lead to a reject of the 'early'
OPEN from modem that made logging fail when activated (it worked
at following boot as MTS opened the log device before the OPEN
was received and this did the channel 4 => TRACE association).

Fix the issue by mapping EDLP channel 4 to TRACE channel at DLP
driver load.

Change-Id: Ia19fb2b1a71838c1f9928d1a339677c0be3158ca
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-4728
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Reviewed-on: https://android.intel.com:443/297129